requirements.txt
bitsandbytes==0.40.2
datasets==2.14.6
einops==0.7.0
ftfy==6.1.3
iopath==0.1.10
numpy==1.24.3
peft==0.4.0
Pillow==10.0.1
Pillow==10.1.0
pytorchvideo==0.1.5
regex==2023.10.3
timm==0.6.7
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.31.0
trl==0.4.7